+++ projects/demo/src/app/types/message.ts
@@ -1,0 +1,3 @@
+export interface Message {
+  readonly text: string;
+}
